// projects/templates/src/lib/components/po-page-change-password/interfaces/po-page-change-password.interface.ts
/**
 * @usedBy PoPageChangePasswordComponent
 *
 * @description
 *
 * Interface com a definição do objeto gerado pelo formulário do componente `po-page-change-password`.
 */
export interface PoPageChangePassword {
  /** Senha atual */
  currentPassword?: string;

  /** Nova senha */
  newPassword?: string;
}